requirements.txt
pytube==15.0.0
streamlit==1.33.0
google-generativeai==0.7.1
pydub==0.25.1
